Module bevy_yoleck::vpeol

Viewport Editing Overlay - utilities for editing entities from a viewport.

This module does not do much, but provide common functionalities for more concrete modules like vpeol_2d and vpeol_3d.

Structs

YoleckKnobClick

YoleckRouteClickTo

Marker for viewport editor overlay plugins to route child interaction to parent entites.

YoleckVpeolSelectionCuePlugin

Add a pulse effect when an entity is being selected.

YoleckWillContainClickableChildren

Marker for entities that will be interacted in the viewport using their children.

Functions

handle_clickable_children_system

Add YoleckRouteClickTo of entities marked with YoleckWillContainClickableChildren.