Module bevy_yoleck::vpeol_2d

Viewport Editing Overlay for 2D games.

Use this module to implement simple 2D editing for 2D games.

To use add the egui and Yoleck plugins to the Bevy app, as well as the plugin of this module:

app.add_plugin(EguiPlugin);
app.add_plugin(YoleckPluginForEditor);
app.add_plugin(YoleckVpeol2dPlugin);

Entity selection by clicking on it is supported by just adding the plugin. To implement dragging, there are two options. Either use the passed data:

app.add_yoleck_handler({
    YoleckTypeHandler::<Example>::new("Example")
        .edit_with(edit_example)
        .populate_with(populate_example)
});

fn edit_example(mut edit: YoleckEdit<Example>) {
    edit.edit(|ctx, data, _ui| {
        if let Some(pos) = ctx.get_passed_data::<Vec2>() {
            data.position = *pos;
        }
    });
}

fn populate_example(mut populate: YoleckPopulate<Example>) {
    populate.populate(|_ctx, data, mut cmd| {
        cmd.insert_bundle(SpriteBundle {
            transform: Transform::from_translation(data.position.extend(0.0)),
            // Actual sprite components
            ..Default::default()
        });
    });
}

Alternatively, use yoleck_vpeol_position_edit_adapter.

Re-exports

pub use crate::vpeol::YoleckWillContainClickableChildren;

Structs

YoleckVpeol2dPlugin

Add the systems required for 2D editing.

YoleckVpeolTransform2dProjection

See yoleck_vpeol_position_edit_adapter.

Functions

yoleck_vpeol_position_edit_adapter

Edit a Vec2 position field of an entity with drag&drop.