Module bevy_yoleck::vpeol_3d

Viewport Editing Overlay for 3D games.

Use this module to implement simple 3D editing for 3D games.

To use add the egui and Yoleck plugins to the Bevy app, as well as the plugin of this module:

app.add_plugin(EguiPlugin);
app.add_plugin(YoleckPluginForEditor);
app.add_plugin(YoleckVpeol3dPlugin);

YoleckVpeol3dPlugin adds the orbit camera plugin from bevy-orbit-controls, but camera still needs to be confiugred for it.

Entity selection by clicking on it is supported by bevy_mod_picking. It’s plugin is added by YoleckVpeol3dPlugin, but the camera and entities still need to be configured.

YoleckVpeol3dCameraBundle can be used to configure the camera for both plugins.

To use vpeol_3d for an entity type, it’s edit system can get Vec3 (and possibly Quat) from the context with get_passed_data. It also needs to add PickableBundle and bevy_transform_gizmo’s GizmoTransformable (both are conveniently re-exported from this module):

app.add_yoleck_handler({
    YoleckTypeHandler::<Example>::new("Example")
        .edit_with(edit_example)
        .populate_with(populate_example)
});

fn edit_example(mut edit: YoleckEdit<Example>) {
    edit.edit(|ctx, data, _ui| {
        if let Some(pos) = ctx.get_passed_data::<Vec3>() {
            data.position = *pos;
        }
        if let Some(rot) = ctx.get_passed_data::<Quat>() {
            data.rotation = *rot;
        }
    });
}

fn populate_example(mut populate: YoleckPopulate<Example>) {
    populate.populate(|_ctx, data, mut cmd| {
        cmd.insert_bundle(PbrBundle {
            transform: Transform::from_translation(data.position),
            // Actual PBR components
            ..Default::default()
        });
        cmd.insert_bundle(PickableBundle::default());
        cmd.insert(GizmoTransformable);
    });
}

3D entities are often hierarchical (e.g. when created by spawn_scene), so PickableBundle needs to be added to the relevant children - together with YoleckRouteClickTo. Since these entities are usually added later, YoleckWillContainClickableChildren can be used:

fn populate_example(mut populate: YoleckPopulate<Example>, asset_server: Res<AssetServer>) {
    populate.populate(|_ctx, data, mut cmd| {
        cmd.insert_bundle(TransformBundle::from_transform(Transform::from_translation(data.position)));
        cmd.with_children(|commands| {
            commands.spawn_scene(asset_server.load("models/my-model.glb#Scene0"));
        });
        cmd.insert(YoleckWillContainClickableChildren);
        cmd.insert(GizmoTransformable); // added on the parent entity, not the children
    });
}

Alternatively, use yoleck_vpeol_transform_edit_adapter. It’ll apply the transform and add PickableBundle and GizmoTransformable, but not YoleckWillContainClickableChildren - that one still needs to be added separately.

Re-exports

pub use crate::vpeol::YoleckWillContainClickableChildren;

Structs

GizmoTransformable

Reexported from bevy_transform_gizmo.

OrbitCameraBundle

Reexported from bevy-orbit-controls.

OrbitCameraController

Reexported from bevy-orbit-controls. A 3rd person camera that orbits around the target.

PickableBundle

Reexported from bevy_mod_picking.

YoleckVpeol3dCameraBundle

Helper for installing bevy-orbit-controls and bevy_mod_picking in a camera entity.

YoleckVpeol3dPlugin

Add the systems required for 2D editing.

YoleckVpeolTransform3dProjection

See yoleck_vpeol_transform_edit_adapter.

Functions

yoleck_vpeol_transform_edit_adapter

Implement parts of the 3D editing for the entity: